interface split
